Add v0.1a1 project files from disk to git repository
Initialize the main project files in the master branch of the git
repository.  In particular, the following files were committed:

-root
    application.py - Program entry point
    .gitignore - Specify which files to ignore for commits
    MANIFEST.in - Tells which directories to include with install
    setup.py - Contains info needed for source distributions

-/vaspcat/src
    poscar.py - Generates POSCAR files from geometry input
    potcar.py - Concatinates POTCAR files in POSCAR atomic order

-/vaspcat/extend
    posext.py - Contains read/parse methods for geometry files

Incar, KPoints, and pbs.in functionality must still be added to
future revisions of the project.  Additionally, some bugfixes
on the poscar.py module to fix Bug #1 must be performed.