Wrap cif_read and cif_parse methods in class
As VaspCat continues to grow, its  modules must be independent of
one another and adaptable to future feature additions.  In this
spirit, the cif_read and cif_parse methods were placed in a new
class Cif inside of posext.py.

Before, if new file formats were added in posext.py, poscar.py
had to be changed before it recognized the new file types.
Now, poscar.py scans posext for the classes contained in posext.
These names coorespond to file extensions, information which
poscar uses when scanning for files to convert.

To accommodate these changes, poscar.py was modified as well,
adding a new attribute 'supported' in main.  This is a list
containing the classes present in posext.py.