Fix Bug #1: Atoms not Appearing
This commit updates poscar.py to properly convert cif files to POSCAR
files if the cif file has extra atoms specified by the
_geom_bond_site_symmetry variable.

Before, the _geom_bond variable was ignored, but it turns out to be
very important.  The variable contains equations specifying how the
fractional positions of extra atoms in the unit cell should be found.
The modifications to poscar.py allow the program to use these
equations.  Before being merged with the master branch, this bugfix will
have to be tested with additional cif files.