Add duplicate coordinate finder to posext.py
In instances where duplicate coordinates are added by the space group
equations, the program now has the capability to remove the duplicates.
Also, the program is able to turn negative fractional
coordinates to positive ones.

A minor bug was also fixed with this commit, in which running VaspCat
in an empty folder crashed the program.  This was accomplished by
changing a line of code in poscar.py having to do with how the program
indicated to the user what file types were supported.