Redo Bugfix #1: Atoms not Appearing
This commit adds a new file, spacegroup.py, originally from the open
source program cif2cell.  It contains a dictionary mapping Hall symbols
to a list of general position equations.  These equations are then used
by posext.py to find the coordinates of every atom in the original cif
file.  Duplicate coordintes are removed by the same method as in the
original bugfix.

The intention of this change is to make how the program finds new atoms
for a given space group more reliable then relying on a bonding
variable that may or may not be found in a given cif file.