Change lattice vector basis equation
Fixed a bug where cif and pdb files saved by VaspCat gave incorrect
basis vectors.  This impacted unit cell shape when the exported POSCAR
file was visualized in a program such as VESTA.

This commit fixes the problem by changing one of the angles in the basis
equation from beta to gamma.  It also switches the equation for
coordinate a13 with that of c31.