Update lattice vector basis equation
Further updates the equations converting cell parameters to unit cell
basis vectors by taking the transpose of the original matrix.  This fix
was tested with a cif file which previously gave the correct coordinates
in the wrong order.

The code contained in this commit fixes the problem, and it is not
anticipated that further updates to the unit cell matrix will be needed.