Modify Duplicate Finder
This commit changes the way duplicate coordinates are found in the Cif
and Pdb classes of filetype.py.  Instead of seeing if the coordinates
are within a certain tolerance value of each other, the coordinates are
rounded to either 3 or 4 decimal places and compared using the set
functionality of Python.

The indices of the rounded duplicates are added to a list, and then used
to remove redundant atomic positions from the unrounded coordinates.
This approach makes it possible to quickly find duplicates in coordinate
sets of over 10,000 atoms, which was previously infeasable.

TODO:
- Correct bug where atom names read from cif or pdb files have extra
  letters as labels.
- Add ability to make atom names the POSCAR file title.
- Create basic INCAR file creation capability.
- Create basic KPOINTS file creation capability.
- Create documentation for settings file and VaspCat modules.