Fix blank POTCAR output bug
This commit deals with an issue users were having where POTCAR files
were empty after being saved.

The issue was in how VaspCat wrote the contents of input POTCAR files to
the output file.  Essentially, the output POTCAR file was repeatedly
opened for writing every time a new input POTCAR file was to be added.
This had the net effect of creating a blank POTCAR file once output was
complete.

As part of the new POTCAR writing code, the module shutil from the
standard library is no longer required to concatinate input POTCAR files.

CHANGELOG:
- Reposition for loops responsible for reading input POTCAR files so
  that their contents are actually saved to the output file.
- Changed how the output POTCAR file is open, using with statements instead
  of separate open statements.
- Removed shutil import statement, as this module is no longer needed.

TODO:
- Modify KPOINTS config file to gamma-centered generation scheme by
  default.  This is because the mesh size required by a particular user
  can't be known ahead of time.
- Fix bug where output folders can't be created if a file is present in
  the working directory with the same name.
- Transition docstrings to Sphinx rst format.
- Document and add boilerplate text to modules missing these docstrings.
- Add additional INCAR options depending on demand by program users.